Move website related files to `src/site` and account related files into `src/features/accounts`
