+++ src/site/errors/WorkInProgress.js
@@ -1,6 +1,6 @@
 import Breadcrumbs from "../../common/components/Breadcrumbs";
 import PageTemplate from "../../common/components/PageTemplate";
-import { Button, VStack, Box, Heading, Text } from "@chakra-ui/react";
+import { Box, Button, Heading, Text, VStack } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
 
 export default function PageNotFound() {
