+++ src/site/user/email-verification/EmailSent.js
